+++ CHANGELOG.md
@@ -80,6 +80,8 @@
   [#10400](https://github.com/Kong/kong/pull/10400)
 - Allow configuring custom error templates
   [#10374](https://github.com/Kong/kong/pull/10374)
+- Avoid compare same table of router's destinations
+  [#10470](https://github.com/Kong/kong/pull/10470)
 
 #### Admin API