+++ CHANGELOG.md
@@ -67,6 +67,14 @@
 
 ## Unreleased
 
+### Fixes
+
+#### Core
+
+- Fix an issue where balancer passive healthcheck would use wrong status code when kong changes status code
+  from upstream in `header_filter` phase.
+  [#10325](https://github.com/Kong/kong/pull/10325)
+
 ## [3.0.0]
 
 > Released 2022/09/12