+++ CHANGELOG.md
@@ -76,7 +76,9 @@
   their consumers fail to process the entries.  Instead, old batches are now dropped
   and an error is logged.
   [#10247](https://github.com/Kong/kong/pull/10247)
-
+- Fix an issue where balancer passive healthcheck would use wrong status code when kong changes status code
+  from upstream in `header_filter` phase.
+  [#10325](https://github.com/Kong/kong/pull/10325)
 
 ## [3.0.1]